Pass breadcrumbs include self flag through request, move and adapt old test
